bot/requirements.txt
aiogram>=3.28,<4.0
SQLAlchemy>=2.0,<3.0
aiosqlite>=0.20,<1.0
httpx>=0.27,<1.0
python-dotenv>=1.0,<2.0
